Interim LLDB support completed
Reverted to GDB command handling logging being commented out. Switched
to the generated register descriptions for the general purpose
registers. This is incomplete work as it doesn't attempt to cover other
register sets (FP, CSR, custom). This was implemented as part of
attempting to get full backtrace support in the LLDB client, which still
does not work (however surprisingly it does work in VSCode)

64-bit GP registers should work but have not been tested.