Update the trace to activate whenever we reach the PC-range and trace is deactive
After further consideration it became apparent that while we need to hook into an instruction in the function prologue to enable trace activation at the time of prologue, it is not of interest at this time what instruction activates the trace. So we only need to consider whether we are within the range and whether the trace has yet been activated.